OBPIH-7866 Extract buildReceivingTableRows to a separate util
